Phase 6d: carve the tie-break flow out of app.js into tiebreak.ts
tiebreak.ts owns the tie-break modal state machine (partner selection ->
sub-round 1 -> partner swap -> final vote -> results) with all 16 state
variables module-private — a reference scan showed nothing outside the
block ever read them. app.js enters via startTiebreakFlow() from the
/api/end_game tiebreak_required branch, after initTiebreakDeps() wires
getSessionId / getInitialOrder / showResults.

types.ts gains the tie-break endpoint contract (TiebreakStartData,
TiebreakSetPartnersResponse, TiebreakAdvanceResponse,
TiebreakVoteResponse + TiebreakRoleResult) mirroring
web/routes/game.py; the fetch calls moved to the typed postJson client.

Verified: build + tsc + eslint clean; 9-step Playwright smoke passed;
targeted browser check forcing a real 2v2 tie and walking the entire
modal (partner selects with confirm gating, both advances, final vote
with submit gating, winner banners, finalize to the results screen) —
zero JS errors.
